custom-charset for Private Use character codepoints: https://github.com/metanorma/metanorma/issues/339:
