+++ lib/metanorma/standoc/converter.rb
@@ -5,6 +5,7 @@ require_relative "front"
 require_relative "lists"
 require_relative "ref"
 require_relative "inline"
+require_relative "anchor"
 require_relative "blocks"
 require_relative "section"
 require_relative "table"
